pyproject.toml
[tool.poetry]
name = "flaskproject"
version = "0.1.0"
description = ""
authors = ["Max Rychel <mrychel@entrustedmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.9"
flask-restful = "^0.3.10"
pyjwt = "^2.8.0"
marshmallow-sqlalchemy = "^1.0.0"
pymongo = "^4.8.0"
mongoengine = "^0.28.2"
bcrypt = "^4.1.3"
flask-jwt-extended = "^4.6.0"
pytest-flask = "^1.3.0"
flake8 = "^7.1.0"
mongomock = "^4.1.2"
setuptools = "^70.2.0"
flask-mongoengine = {git = "https://github.com/idoshr/flask-mongoengine.git", rev = "1.0.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
minversion = "8.2.2"
addopts = "-ra -q"
testpaths = [
    "tests.py"
]
